// src/app/generated_services/model/createCompanyPersonDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCompanyPersonDTO { 
    personId?: string;
    email: string;
    phoneNumber?: string | null;
    name: string | null;
    foundedIn?: string | null;
    cnpj: string | null;
}