// src/app/generated_services/model/createIndividualPersonDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateIndividualPersonDTO { 
    personId?: string;
    email: string;
    phoneNumber?: string | null;
    photoUrl?: string | null;
    firstName: string | null;
    lastName: string | null;
    cpf: string | null;
}